Fix float precision

+++ examples/datum/Arduino/rx_datum/rx_datum.ino
@@ -3,7 +3,7 @@
 
 SerialTransfer myTransfer;
 
-float y;
+double y;
 
 
 void setup()
